Update DetectLM.py
Kestrel analytics cannot change a view which is resulted from applying the SORT transformer on a variable. Thus, I set the first_observed column of the dataframe representing observations to the column of the other two dataframes, obtained from parguet files, which shows the time of observations. Also, I edited the first line of function "process" which was causing error.